Add files via upload
create an “Online Inventory  System” for MIC. For this, you have to create a MIC_Inventory class that   
includes the following methods: 
DisplayEquipment() : To display the available equipment (for example Motor,IR Module,Arduino uno etc)
LendEquipment(): To lend a equipment to a student
AddEquipment(): To add a equipment to the MIC_inventory
ReturnEquipment(): To return the equipment to the MIC_invenbtory. 
As you have created a MIC_inventory class ow you will create an object and pass them within a list in this method: MIC_invetory_2022= MIC_invetory (['Arduino_uno', 'Motor', 'Ir_module', 'ECG sensor'',])
then, create a main function and run an infinite while loop that asks the students for their input that whether they want to display, lend, add or return a equipment.